Make TestNavigationService PanelNavigationService
